+++ web/modules/custom/ai_screening_project/src/Helper/ProjectHelper.php
@@ -11,8 +11,11 @@ use Drupal\Core\Field\BaseFieldDefinition;
 use Drupal\Core\Logger\LoggerChannel;
 use Drupal\Core\Session\AccountProxyInterface;
 use Drupal\Core\StringTranslation\StringTranslationTrait;
+use Drupal\ai_screening\Helper\AbstractHelper;
+use Drupal\ai_screening_project_track\Helper\ProjectTrackHelper;
 use Drupal\ai_screening_project_track\ProjectTrackStatus;
 use Drupal\ai_screening_project_track\ProjectTrackStorageInterface;
+use Drupal\ai_screening_project_track\ProjectTrackToolStorageInterface;
 use Drupal\core_event_dispatcher\CoreHookEvents;
 use Drupal\core_event_dispatcher\EntityHookEvents;
 use Drupal\core_event_dispatcher\Event\Core\CronEvent;
@@ -20,6 +23,7 @@ use Drupal\core_event_dispatcher\Event\Entity\EntityAccessEvent;
 use Drupal\core_event_dispatcher\Event\Entity\EntityBaseFieldInfoEvent;
 use Drupal\core_event_dispatcher\Event\Entity\EntityDeleteEvent;
 use Drupal\core_event_dispatcher\Event\Entity\EntityInsertEvent;
+use Drupal\group\Entity\GroupInterface;
 use Drupal\group\Entity\GroupRelationshipInterface;
 use Drupal\group\Entity\Storage\GroupRelationshipStorageInterface;
 use Drupal\group\Entity\Storage\GroupStorage;
@@ -29,7 +33,6 @@ use Drupal\preprocess_event_dispatcher\Event\NodePreprocessEvent;
 use Drupal\taxonomy\TermStorageInterface;
 use Drupal\user\UserStorageInterface;
 use Drupal\webform\WebformSubmissionStorageInterface;
-use Psr\Log\LoggerAwareInterface;
 use Psr\Log\LoggerAwareTrait;
 use Psr\Log\LoggerTrait;
 use Symfony\Component\EventDispatcher\EventSubscriberInterface;
@@ -38,7 +41,7 @@ use Symfony\Component\HttpKernel\Exception\AccessDeniedHttpException;
 /**
  * A helper class for the Project node entity.
  */
-class ProjectHelper implements LoggerAwareInterface, EventSubscriberInterface {
+class ProjectHelper extends AbstractHelper implements EventSubscriberInterface {
   use LoggerAwareTrait;
   use LoggerTrait;
   use StringTranslationTrait;
@@ -97,14 +100,22 @@ class ProjectHelper implements LoggerAwareInterface, EventSubscriberInterface {
   private readonly ProjectTrackStorageInterface|EntityStorageInterface $projectTrackStorage;
 
   /**
+   * The project track tool storage.
+   *
+   * @var \Drupal\ai_screening_project_track\ProjectTrackToolStorageInterface|\Drupal\Core\Entity\EntityStorageInterface
+   */
+  private readonly ProjectTrackToolStorageInterface|EntityStorageInterface $projectTrackToolStorage;
+
+  /**
    * Constructor.
    */
   public function __construct(
     private readonly AccountProxyInterface $accountProxy,
+    private readonly ProjectTrackHelper $projectTrackHelper,
     EntityTypeManagerInterface $entityTypeManager,
     LoggerChannel $logger,
   ) {
-    $this->setLogger($logger);
+    parent::__construct($logger);
     $this->groupStorage = $entityTypeManager->getStorage('group');
     $this->groupRelationshipStorage = $entityTypeManager->getStorage('group_relationship');
     $this->userStorage = $entityTypeManager->getStorage('user');
@@ -112,6 +123,7 @@ class ProjectHelper implements LoggerAwareInterface, EventSubscriberInterface {
     $this->termStorage = $entityTypeManager->getStorage('taxonomy_term');
     $this->webformSubmissionStorage = $entityTypeManager->getStorage('webform_submission');
     $this->projectTrackStorage = $entityTypeManager->getStorage('project_track');
+    $this->projectTrackToolStorage = $entityTypeManager->getStorage('project_track_tool');
   }
 
   /**
@@ -132,41 +144,15 @@ class ProjectHelper implements LoggerAwareInterface, EventSubscriberInterface {
       $relationships = $this->groupRelationshipStorage->loadMultiple($relationshipIds);
 
       $groups = $this->groupStorage->loadMultiple(array_map(
-          static fn (GroupRelationshipInterface $relationship) => $relationship->getGroupId(),
-          $relationships)
-      );
+        static fn (GroupRelationshipInterface $relationship) => $relationship->getGroupId(),
+        $relationships
+      ));
       foreach ($groups as $group) {
         $group->delete();
       }
 
-      // Get all project tracks for the project.
-      $projectTrackIds = $this->projectTrackStorage->getQuery()
-        ->accessCheck(FALSE)
-        ->condition('project_id', $project->id(), '=')
-        ->execute();
-
-      $submissionIds = [];
-      foreach ($projectTrackIds as $projectTrackId) {
-        // Get all webforms that reference these project tracks.
-        $submissionIds = $this->webformSubmissionStorage->getQuery()
-          ->accessCheck(FALSE)
-          ->condition('entity_type', 'project_track', '=')
-          ->condition('entity_id', $projectTrackId, '=')
-          ->execute();
-      }
-
-      $projectTracks = $this->projectTrackStorage->loadMultiple($projectTrackIds);
-      $webformSubmissions = $this->webformSubmissionStorage->loadMultiple($submissionIds);
-
-      // Delete project tracks and webform submissions for the project.
-      foreach ($webformSubmissions as $webformSubmission) {
-        $webformSubmission->delete();
-      }
-
-      foreach ($projectTracks as $projectTrack) {
-        $projectTrack->delete();
-      }
-
+      $projectTracks = $this->loadProjectTracks($project);
+      $this->projectTrackHelper->deleteProjectTracks($projectTracks);
     }
     catch (\Exception $exception) {
       $this->error('Error deleting project: @message', [
@@ -204,9 +190,12 @@ class ProjectHelper implements LoggerAwareInterface, EventSubscriberInterface {
   public function entityAccess(EntityAccessEvent $event): void {
     $entity = $event->getEntity();
 
-    // Deny access if content is corrupted.
-    if ($this->isCorrupted($entity)) {
-      $event->setAccessResult(AccessResult::forbidden(sprintf('Entity %s (%s) is corrupted', $entity->label(), $entity->id())));
+    if ($entity instanceof NodeInterface) {
+      // Deny access if content is corrupted.
+      if ($this->isCorrupted($entity)) {
+        $event->setAccessResult(AccessResult::forbidden(sprintf('Entity %s (%s) is corrupted', $entity->label(),
+          $entity->id())));
+      }
     }
   }
 
@@ -216,6 +205,7 @@ class ProjectHelper implements LoggerAwareInterface, EventSubscriberInterface {
   public function entityInsert(EntityInsertEvent $event): void {
     $entity = $event->getEntity();
     if ($this->isProject($entity)) {
+      assert($entity instanceof NodeInterface);
       $this->addProjectGroup($entity);
       $this->addProjectTracks($entity);
     }
@@ -347,37 +337,46 @@ class ProjectHelper implements LoggerAwareInterface, EventSubscriberInterface {
       // Add project tracks to project.
       $projectTrackTerms = $this->termStorage->loadMultiple($projectTrackTermIds);
 
+      $projectTrackCounter = 0;
       foreach ($projectTrackTerms as $projectTrackTerm) {
-        /** @var \Drupal\taxonomy\TermInterface $projectTrackTerm */
-        $webformId = $projectTrackTerm->field_webform->target_id;
-        $webformSubmission = $this->webformSubmissionStorage->create([
-          'webform_id' => $webformId,
-          'entity_type' => 'project_track',
-        ]);
-
-        $webformSubmission->save();
-        $submissionId = $webformSubmission->id();
-
         $projectTrack = $this->projectTrackStorage
           ->create([
             'type' => $projectTrackTerm->id(),
             'title' => $projectTrackTerm->getName(),
             'description' => $projectTrackTerm->getDescription(),
             'project_track_evaluation' => '0',
-            'project_id' => $entity->id(),
-            'tool_id' => $submissionId,
-            'tool_entity_type' => 'webform_submission',
+            'project_id' => $entity,
           ])
-          ->setProjectTrackStatus(ProjectTrackStatus::NEW);
-
+          ->setProjectTrackStatus(ProjectTrackStatus::NEW)
+          ->setDelta($projectTrackCounter++);
         $projectTrack->save();
-        $projectTrackId = $projectTrack->id();
 
-        /** @var \Drupal\webform\WebformSubmissionInterface $submission */
-        $submission = $this->webformSubmissionStorage->load($submissionId);
-        $submission->set('entity_id', $projectTrackId);
+        /** @var \Drupal\webform\WebformInterface[] $webforms */
+        $webforms = $projectTrackTerm->get('field_webform')->referencedEntities();
 
-        $submission->save();
+        $toolCounter = 0;
+        foreach ($webforms as $webform) {
+          $tool = $this->projectTrackToolStorage->create([
+            'project_track_id' => $projectTrack->id(),
+            'tool_entity_type' => 'webform_submission',
+          ]);
+          $tool->setDelta($toolCounter++);
+          $tool->save();
+
+          $webformSubmission = $this->webformSubmissionStorage->create([
+            'webform' => $webform,
+            'entity_type' => 'project_track_tool',
+          ]);
+          $webformSubmission->save();
+
+          // Reload the webform submission.
+          $webformSubmission = $this->webformSubmissionStorage->load($webformSubmission->id());
+          $webformSubmission->set('entity_id', $tool->id());
+          $webformSubmission->save();
+
+          $tool->set('tool_id', $webformSubmission->id());
+          $tool->save();
+        }
       }
     }
     catch (\Exception $exception) {
@@ -414,24 +413,52 @@ class ProjectHelper implements LoggerAwareInterface, EventSubscriberInterface {
    */
   public function preprocessProject(NodePreprocessEvent $event): void {
     $variables = $event->getVariables();
-    $projectTrackIds = $this->projectTrackStorage->getQuery()
+    $node = $variables->getEntity();
+    if ($node instanceof NodeInterface) {
+      $variables->set('projectTracks', $this->loadProjectTracks($node));
+      $variables->set('projectGroup', $this->loadProjectGroup($node));
+      $variables->set('projectMembers', $this->loadProjectGroup($node)->getMembers());
+    }
+  }
+
+  /**
+   * Load project tracks.
+   *
+   * @param \Drupal\node\NodeInterface $project
+   *   The project.
+   *
+   * @return \Drupal\ai_screening_project_track\ProjectTrackInterface[]
+   *   The tracks.
+   */
+  public function loadProjectTracks(NodeInterface $project): array {
+    $ids = $this->projectTrackStorage->getQuery()
       ->accessCheck(FALSE)
-      ->condition('project_id', $variables->getEntity()->id(), '=')
+      ->condition('project_id', $project->id(), '=')
+      ->sort('delta')
       ->execute();
 
-    $variables->set('projectTracks', $this->projectTrackStorage->loadMultiple($projectTrackIds));
+    return $this->projectTrackStorage->loadMultiple($ids);
+  }
 
+  /**
+   * Load project group.
+   *
+   * @param \Drupal\node\NodeInterface $project
+   *   The project.
+   *
+   * @return \Drupal\group\Entity\GroupInterface
+   *   The group
+   */
+  public function loadProjectGroup(NodeInterface $project) : GroupInterface {
     $relationshipIds = $this->groupRelationshipStorage->getQuery()
       ->accessCheck(FALSE)
-      ->condition('entity_id', $variables->getEntity()->id(), '=')
+      ->condition('entity_id', $project->id(), '=')
       ->condition('type', 'project_group-group_node-project', '=')
       ->execute();
 
     $relationships = $this->groupRelationshipStorage->loadMultiple($relationshipIds);
-    $group = $this->groupStorage->load(reset($relationships)->getGroupId());
 
-    $variables->set('projectGroup', $group);
-    $variables->set('projectMembers', $group->getMembers());
+    return $this->groupStorage->load(reset($relationships)->getGroupId());
   }
 
 }