+++ web/modules/custom/ai_screening_project_track/src/Entity/ProjectTrack.php
@@ -4,7 +4,6 @@ declare(strict_types=1);
 
 namespace Drupal\ai_screening_project_track\Entity;
 
-use Drupal\Core\Datetime\DrupalDateTime;
 use Drupal\Core\Entity\EntityChangedTrait;
 use Drupal\Core\Entity\EntityTypeInterface;
 use Drupal\Core\Entity\RevisionableContentEntityBase;
@@ -12,8 +11,6 @@ use Drupal\Core\Field\BaseFieldDefinition;
 use Drupal\ai_screening_project_track\ProjectTrackInterface;
 use Drupal\ai_screening_project_track\ProjectTrackStatus;
 use Drupal\node\NodeInterface;
-use function Safe\json_decode;
-use function Safe\json_encode;
 
 /**
  * Defines the project track entity class.
@@ -49,7 +46,6 @@ use function Safe\json_encode;
  *   base_table = "project_track",
  *   revision_table = "project_track_revision",
  *   show_revision_ui = TRUE,
- *   admin_permission = "administer project_track",
  *   entity_keys = {
  *     "id" = "id",
  *     "revision" = "revision_id",
@@ -65,7 +61,7 @@ use function Safe\json_encode;
  *     "collection" = "/admin/content/project-track",
  *     "add-form" = "/project-track/add",
  *     "canonical" = "/project-track/{project_track}",
- *     "edit-form" = "/project-track/{project_track}",
+ *     "edit-form" = "/project-track/{project_track}/edit",
  *     "delete-form" = "/project-track/{project_track}/delete",
  *     "delete-multiple-form" = "/admin/content/project-track/delete-multiple",
  *     "revision" = "/project-track/{project_track}/revision/{project_track_revision}/view",
@@ -78,6 +74,8 @@ use function Safe\json_encode;
 final class ProjectTrack extends RevisionableContentEntityBase implements ProjectTrackInterface {
 
   use EntityChangedTrait;
+  use SortableEntityTrait;
+  use TimestampableEntityTrait;
 
   /**
    * {@inheritdoc}
@@ -110,7 +108,7 @@ final class ProjectTrack extends RevisionableContentEntityBase implements Projec
 
     $fields['project_track_status'] = BaseFieldDefinition::create('string')
       ->setLabel(t('Project track status'))
-      ->setDescription(t('The type of the project track.'));
+      ->setDescription(t('The status of the project track.'));
 
     $fields['created'] = BaseFieldDefinition::create('created')
       ->setLabel(t('Authored on'))
@@ -125,18 +123,8 @@ final class ProjectTrack extends RevisionableContentEntityBase implements Projec
       ->setSetting('target_type', 'node')
       ->setSetting('handler_settings', ['target_bundles' => ['project' => 'project']]);
 
-    $fields['tool_id'] = BaseFieldDefinition::create('string')
-      ->setLabel(t('Tool'))
-      ->setDescription(t('The ID of the tool referenced.'));
-
-    $fields['tool_entity_type'] = BaseFieldDefinition::create('string')
-      ->setLabel(t('Tool entity type'))
-      ->setDescription(t('The entity type of the tool referenced.'));
-
-    $fields['tool_data'] = BaseFieldDefinition::create('string_long')
-      ->setLabel(t('Tool data'))
-      ->setRevisionable(TRUE)
-      ->setDescription(t('The data matching the tool configuration'));
+    $fields['delta'] = BaseFieldDefinition::create('integer')
+      ->setLabel(t('Dalte'));
 
     return $fields;
   }
@@ -195,74 +183,10 @@ final class ProjectTrack extends RevisionableContentEntityBase implements Projec
   /**
    * {@inheritdoc}
    */
-  public function getCreated(): DrupalDateTime {
-    return $this->getDateTime('created');
-  }
-
-  /**
-   * {@inheritdoc}
-   */
-  public function getChanged(): DrupalDateTime {
-    return $this->getDateTime('changed');
-  }
-
-  /**
-   * {@inheritdoc}
-   */
   public function getProject(): NodeInterface {
     $entities = $this->get('project_id')->referencedEntities();
 
     return reset($entities);
-  }
-
-  /**
-   * {@inheritdoc}
-   */
-  public function getToolId(): int|string {
-    return $this->get('tool_id')->getString();
-  }
-
-  /**
-   * {@inheritdoc}
-   */
-
-  /**
-   * {@inheritdoc}
-   */
-  public function getToolEntityType(): string {
-    return $this->get('tool_entity_type')->getString();
-  }
-
-  /**
-   * {@inheritdoc}
-   */
-  public function getToolData(): array {
-    $value = $this->get('tool_data')->getString();
-
-    try {
-      return json_decode($value, TRUE);
-    }
-    catch (\Exception $exception) {
-      return [];
-    }
-  }
-
-  /**
-   * {@inheritdoc}
-   */
-  public function setToolData(array $data): self {
-    $this->set('tool_data', json_encode($data));
-
-    return $this;
-  }
-
-  /**
-   * Get datatime from field value.
-   */
-  private function getDateTime(string $field): DrupalDateTime {
-    $value = (int) $this->get($field)->getString();
-
-    return DrupalDateTime::createFromTimestamp($value);
   }
 
 }