+++ web/modules/custom/ai_screening_project_track/src/ProjectTrackStorage.php
@@ -5,10 +5,10 @@ namespace Drupal\ai_screening_project_track;
 use Drupal\Core\Entity\Sql\SqlContentEntityStorage;
 
 /**
- * Defines the storage handler class for nodes.
+ * Defines the storage handler class for project track entities.
  *
  * This extends the base storage class, adding required special handling for
- * node entities.
+ * project track entities.
  */
 class ProjectTrackStorage extends SqlContentEntityStorage implements ProjectTrackStorageInterface {